fix unlink_tmp

+++ minishell/srcs/executer/unlink_heredoc_tmp_file.c
@@ -6,7 +6,7 @@
 /*   By: ccartman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 21:11:03 by ccartman          #+#    #+#             */
-/*   Updated: 2022/02/14 21:24:50 by ccartman         ###   ########.fr       */
+/*   Updated: 2022/02/20 14:58:21 by ccartman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@ int	unlink_heredoc_tmp_file(t_command **cmd)
 {
 	size_t	i;
 	size_t	j;
+	char	*fn;
 
 	i = -1;
 	while (cmd[++i])
@@ -25,8 +26,12 @@ int	unlink_heredoc_tmp_file(t_command **cmd)
 		{
 			if (cmd[i]->redirect[j]->type & REDIRECT_HEREDOC)
 			{
-				if (unlink(cmd[i]->redirect[j]->name) == -1)
-					print_str_error(cmd[i]->redirect[j]->name);
+				fn = cmd[i]->redirect[j]->name;
+				if (!ft_strncmp(fn, HRDC_TMPFILE, sizeof(HRDC_TMPFILE) - 1))
+				{
+					if (unlink(cmd[i]->redirect[j]->name) == -1)
+						print_str_error(cmd[i]->redirect[j]->name);
+				}
 			}
 		}
 	}
